Refactor mqtt schema's to new module
