+++ homeassistant/components/mqtt/light/schema_template.py
@@ -45,7 +45,7 @@ from ..const import (
     PAYLOAD_NONE,
 )
 from ..debug_info import log_messages
-from ..mixins import MQTT_ENTITY_COMMON_SCHEMA, MqttEntity, write_state_on_attr_change
+from ..mixins import MqttEntity, write_state_on_attr_change
 from ..models import (
     MqttCommandTemplate,
     MqttValueTemplate,
@@ -53,6 +53,7 @@ from ..models import (
     ReceiveMessage,
     ReceivePayloadType,
 )
+from ..schemas import MQTT_ENTITY_COMMON_SCHEMA
 from .schema import MQTT_LIGHT_SCHEMA_SCHEMA
 from .schema_basic import MQTT_LIGHT_ATTRIBUTES_BLOCKED
