+++ homeassistant/components/mqtt/tag.py
@@ -20,7 +20,6 @@ from .config import MQTT_BASE_SCHEMA
 from .const import ATTR_DISCOVERY_HASH, CONF_QOS, CONF_TOPIC
 from .discovery import MQTTDiscoveryPayload
 from .mixins import (
-    MQTT_ENTITY_DEVICE_INFO_SCHEMA,
     MqttDiscoveryDeviceUpdate,
     async_handle_schema_error,
     async_setup_non_entity_entry_helper,
@@ -34,6 +33,7 @@ from .models import (
     ReceiveMessage,
     ReceivePayloadType,
 )
+from .schemas import MQTT_ENTITY_DEVICE_INFO_SCHEMA
 from .subscription import EntitySubscription
 from .util import valid_subscribe_topic
